refactor(pagination): move the pagination contracts to shared
The pagination query DTO, response envelope, and metadata builder carry no
domain knowledge, and core/ is for how the application operates rather than
for reusable HTTP building blocks.

They are not moved into features/users, their only current consumer, because
this is a starter template: the second domain that paginates would have to
import from another feature or duplicate them. shared/ is where a genuinely
domain-agnostic block that several domains will consume belongs.

Update AGENTS.md and docs/EDD.md, which also still described the global guard
chain as owned by AccessControlModule and the @Public marker as living in the
auth feature.
