+++ src/features/users/dto/paginated-users-response.dto.ts
@@ -1,4 +1,4 @@
-import { PaginatedResponseDto } from '../../../core/pagination/pagination.dto';
+import { PaginatedResponseDto } from '../../../shared/pagination/pagination.dto';
 import { UserResponseDto } from './user-response.dto';
 
 export class PaginatedUsersResponseDto extends PaginatedResponseDto(
